+++ db/migrate/20230405172134_add_counter_cache_to_developer_conversations.rb
@@ -1,0 +1,7 @@
+class AddCounterCacheToDeveloperConversations < ActiveRecord::Migration[7.0]
+  def change
+    add_column :developers, :conversations_count, :integer, null: false, default: 0
+
+    Developer.find_each { |d| Developer.reset_counters(d.id, :conversations) }
+  end
+end