+++ test/system/filters_test.rb
@@ -1,24 +1,22 @@
 require "application_system_test_case"
 
 class FiltersTest < ApplicationSystemTestCase
-  test "developers availability sort click adds sort param" do
+  test "recommended sort click adds sort param" do
     visit developers_path
-    find(:css, "#sort button[type='button']").click
-    find(:css, "#sort button[type='submit'][value='availability']").click
-
-    assert_current_path(/sort=availability/)
+    sort_by "recommended"
+    assert_current_path(/sort=recommended/)
   end
 
   test "developers newest sort click adds sort param" do
     visit developers_path
-    click_sort_by_newset
+    sort_by "newest"
     assert_current_path(/sort=newest/)
   end
 
   test "developers sort order is persisted with work preferences filters" do
     visit developers_path
     find(:css, "[name='include_not_interested']").set(true)
-    click_sort_by_newset
+    sort_by "newest"
 
     assert_current_path(/sort=newest/)
     assert_current_path(/include_not_interested=1/)
@@ -26,27 +24,25 @@ class FiltersTest < ApplicationSystemTestCase
 
   test "applying multiple filters to developers" do
     visit developers_path
+    sort_by "recommended"
 
-    find(:css, "#sort button[type='button']").click
-    find(:css, "#sort button[type='submit'][value='availability']").click
-
-    assert_current_path(/sort=availability/)
+    assert_current_path(/sort=recommended/)
 
     find(:css, "[name='include_not_interested']").set(true)
     find(:css, "input[type=submit]").click
 
-    assert_current_path(/sort=availability/)
+    assert_current_path(/sort=recommended/)
     assert_current_path(/include_not_interested=1/)
 
-    click_sort_by_newset
+    sort_by "newest"
 
     assert_current_path(/sort=newest/)
     assert_current_path(/include_not_interested=1/)
   end
 
-  def click_sort_by_newset
+  def sort_by(sort)
     toggle_sort_dropdown
-    find(:css, "#sort button[type='submit'][value='newest']").click
+    find(:css, "#sort button[type='submit'][value='#{sort}']").click
   end
 
   def toggle_sort_dropdown