+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### April
 
 * April 11 - Convert hiring agreement to in-app DocuSign #835 @MikeRayUX
+* April 8 - Send subscription renewal email #860 @jkostolansky
+* April 7 - Launch recommended sorting option #859
 * April 5 - Distinguish between new profiles and recently updated ones #856
 * April 4 - Remove `developers.available_on` #854